Update info file when paths are changed on resource
Previous code had a bug in that when the resource changed the paths for
`cert_chain_file`, `chain_file` or `key_file`, the new files would be
written out, but the info file would not reflect the changed paths. This
caused drift which induced a corrective change on every puppet run. The
only workaround previously was to manually edit or remove the info json
file so that the provider would write out a new correct copy.

This change updates the current paths from the resource into the json
file whenever it is flushed. The info file is only rewritten when
changed so that the modification timestamps on the info file match with
when another change was actually made, and to prevent redundant updates
to the file on disk.